refactor: 컴포넌트에서 커스텀훅 분리
