+++ src/hooks/useCartNavigate.ts
@@ -1,0 +1,14 @@
+import { useNavigate } from 'react-router-dom';
+import ROUTES from '../constants/routes';
+
+const useCartNavigate = () => {
+  const navigate = useNavigate();
+
+  const handleHomeButtonClick = () => {
+    navigate(ROUTES.CART);
+  };
+
+  return { handleHomeButtonClick };
+};
+
+export default useCartNavigate;
