+++ src/pages/cart/CartPage.tsx
@@ -1,34 +1,19 @@
 import Header from '../../components/Header/Header';
 import Button from '../../components/common/Button';
-import { useNavigate } from 'react-router-dom';
-import { totalOrderAmountState, totalProductQuantity } from '../../store/selectors';
 import { useRecoilValue } from 'recoil';
 import CartList from './components/CartList';
 import CartTitle from './components/CartTitle';
 import CartTotals from './components/CartTotals';
-import ROUTES from '../../constants/routes';
+import useCartNavigate from '../../hooks/useCartNavigate';
+import useCheckoutNavigate from '../../hooks/useCheckoutNavigate';
 
 import styles from './Cart.module.css';
 import { productsState } from '../../store/atoms';
 
 export default function CartPage() {
-  const navigate = useNavigate();
-  const handleHomeButtonClick = () => {
-    navigate(ROUTES.CART);
-  };
   const products = useRecoilValue(productsState);
-  const { totalCount, totalQuantity } = useRecoilValue(totalProductQuantity);
-  const { totalAmount } = useRecoilValue(totalOrderAmountState);
-
-  const handleFooterButtonClick = () => {
-    navigate(ROUTES.CHECK_OUT, {
-      state: {
-        totalCount,
-        totalQuantity,
-        totalAmount,
-      },
-    });
-  };
+  const { handleFooterButtonClick } = useCheckoutNavigate();
+  const { handleHomeButtonClick } = useCartNavigate();
 
   return (
     <>
