+++ src/store/selectors.ts
@@ -77,15 +77,3 @@ export const totalProductQuantity = selector({
     };
   },
 });
-
-// const isCheckedState = selectorFamily({
-//   key: 'isCheckedState',
-//   get: (id) => ({ get }) => {
-//     const storedValue = window.localStorage.getItem(JSON.stringify(id));
-//     return storedValue !== null ? JSON.parse(storedValue) : false;
-//   },
-//   set: (id) => ({ set }, newValue) => {
-//     set(isCheckedState(id), newValue);
-//     window.localStorage.setItem(JSON.stringify(id), JSON.stringify(newValue));
-//   },
-// });
